moving cluster_integration_test to server package; removing empty "testintegration" package
