--- testintegration/logger.go
@@ -1,10 +1,0 @@
-package testintegration
-
-import (
-	log "github.com/Sirupsen/logrus"
-)
-
-var logger = log.WithFields(log.Fields{
-	"app":    "integrationtest",
-	"module": "cluster",
-})
